chore(producer): more rabbit code and greptile fixes, refactor producer workflows to keep steps and workflows separate
