+++ apps/medusa-be/src/admin/lib/format-locale-code.ts
@@ -1,2 +1,2 @@
 export const formatLocaleCode = (code: string | undefined) =>
-  code?.replace(/([a-z])([A-Z])/g, "$1-$2").replace(/_/g, "-")
+  (code ?? "").replace(/([a-z])([A-Z])/g, "$1-$2").replace(/_/g, "-")
