+++ apps/medusa-be/src/workflows/producer/workflows/index.ts
@@ -1,0 +1,9 @@
+export * from "./create-producer-attribute-types"
+export * from "./create-producers"
+export * from "./delete-producer-attribute-types"
+export * from "./delete-producers"
+export * from "./restore-producer-attribute-types"
+export * from "./restore-producers"
+export * from "./set-producer-products"
+export * from "./set-product-producers"
+export * from "./update-producers"
